refactor: move accordion to ui, move ui one level up
