+++ web_ui/src/pages/project-details/components/project-models/train-model-dialog/advanced-settings/data-management/removing-duplicated/removing-duplicates.component.tsx
@@ -6,7 +6,7 @@ import { FC, useState } from 'react';
 import { Content, ContextualHelp, Flex, Text } from '@adobe/react-spectrum';
 
 import { Switch } from '../../../../../../../../shared/components/switch/switch.component';
-import { Accordion } from '../../accordion/accordion.component';
+import { Accordion } from '../../ui/accordion/accordion.component';
 
 const RemovingDuplicatesTooltip: FC = () => {
     return (
