+++ web_ui/src/pages/project-details/components/project-models/train-model-dialog/advanced-settings/ui/accordion/accordion.component.tsx
@@ -6,7 +6,7 @@ import { ComponentProps, FC, ReactNode } from 'react';
 import { Disclosure, DisclosurePanel, DisclosureTitle, Divider, Flex, Text, View } from '@adobe/react-spectrum';
 import clsx from 'clsx';
 
-import { AlertOutlined } from '../../../../../../../assets/icons';
+import { AlertOutlined } from '../../../../../../../../assets/icons';
 
 import styles from './accordion.module.scss';
