+++ web_ui/src/pages/project-details/components/project-models/train-model-dialog/advanced-settings/ui/accordion/accordion.module.scss
