+++ web_ui/src/pages/project-details/components/project-models/train-model-dialog/advanced-settings/ui/number-parameter.component.tsx
@@ -5,8 +5,8 @@ import { FC, useEffect, useState } from 'react';
 
 import { Flex, NumberField } from '@adobe/react-spectrum';
 
-import { Slider } from '../../../../../../../../shared/components/slider/slider.component';
-import { getFloatingPointStep } from '../../utils';
+import { Slider } from '../../../../../../../shared/components/slider/slider.component';
+import { getFloatingPointStep } from '../utils';
 
 interface NumberGroupParamsProps {
     value: number;
