+++ web_ui/src/pages/project-details/components/project-models/train-model-dialog/advanced-settings/ui/options-buttons/options-buttons.component.tsx
@@ -4,7 +4,7 @@
 import { Flex } from '@adobe/react-spectrum';
 import clsx from 'clsx';
 
-import { Button } from '../../../../../../../../../shared/components/button/button.component';
+import { Button } from '../../../../../../../../shared/components/button/button.component';
 
 import styles from './options-buttons.module.scss';
