+++ web_ui/src/pages/project-details/components/project-models/train-model-dialog/advanced-settings/ui/reset-button.component.tsx
@@ -3,8 +3,8 @@
 
 import { ComponentProps, FC } from 'react';
 
-import { Refresh } from '../../../../../../../../assets/icons';
-import { ActionButton } from '../../../../../../../../shared/components/button/button.component';
+import { Refresh } from '../../../../../../../assets/icons';
+import { ActionButton } from '../../../../../../../shared/components/button/button.component';
 
 type ResetButtonProps = Omit<ComponentProps<typeof ActionButton>, 'children' | 'isQuiet'>;
